pyproject.toml
[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
name = "root"
version = "0.0.0"
[tool.poetry.dependencies]
"discord.py" = "*"
flask = "*"
python = "^3.7"
requests = "*"
time = "*"
